Updating header file

+++ 0x14-bit_manipulation/holberton.h
@@ -6,7 +6,7 @@
 #include <unistd.h>
 
 unsigned int binary_to_uint(const char *b);
-int _pow(int a);
+double _pow(double a);
 void print_binary(unsigned long int n);
 int get_bit(unsigned long int n, unsigned int index);
 int set_bit(unsigned long int *n, unsigned int index);
